runner/utils/config.py
"""
配置加载工具
"""

from pathlib import Path
import yaml


def load_config() -> dict:
    config_path = Path(__file__).resolve().parent.parent.parent / "config.yaml"
    if not config_path.exists():
        raise FileNotFoundError(f"Config not found: {config_path}")
    with open(config_path, "r", encoding="utf-8") as f:
        return yaml.safe_load(f)
